ai_backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
Pillow==10.4.0
opencv-python==4.9.0.80
numpy<2.0.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
git+https://github.com/openai/CLIP.git
# AI Model Dependencies
ftfy
regex
tqdm
